backend/requirements.txt
# requirements.txt
requests
pandas
pyodbc
tqdm
boto3
python-dotenv
